+++ CHANGELOG.md
@@ -5,3 +5,9 @@ All notable changes to `laravel-mapbox` will be documented in this file.
 ## 0.0.1 - 2021-29-11
 
 - initial release
+
+## 0.0.2 - 2021-13-12
+- Added RTL Support for the names of places 
+- Added support to cooperative gestures
+- Added free marker support, so the end user can interact with the map
+